+++ src/Aardvark.Base.FSharp/Utilities/Native.fs
@@ -192,6 +192,38 @@ module NativeUtilities =
     let inline pinArr ([<InlineIfLambda>] f: nativeptr<'T> -> 'U) (array: 'T[])  =
         NativePtr.pinArr f array
 
+    /// Utility to pin values with IDisposable semantics.
+    type PinnedValue private (value: obj, length: int) =
+        let gc = GCHandle.Alloc(value, GCHandleType.Pinned)
+        let address = gc.AddrOfPinnedObject()
+
+        new (value: obj) =
+            let length = match value with | :? Array as array -> array.Length | _ -> 1
+            new PinnedValue(value, length)
+
+        new (array: Array) =
+            new PinnedValue(array, array.Length)
+
+        /// The address of the pinned value.
+        member _.Address = address
+
+        /// The number of elements if the pinned value is an array, 1 otherwise.
+        member _.Length = length
+
+        member _.Dispose() = gc.Free()
+
+        interface IDisposable with
+            member this.Dispose() = this.Dispose()
+
+    /// Utility to pin values with IDisposable semantics.
+    type PinnedValue<'T when 'T : unmanaged> =
+        inherit PinnedValue
+        new (value: 'T) = { inherit PinnedValue(value :> obj) }
+        new (array: 'T[]) = { inherit PinnedValue(array :> Array) }
+
+        /// The pointer of the pinned value.
+        member inline this.Pointer : nativeptr<'T> = NativePtr.ofNativeInt this.Address
+
 [<AutoOpen>]
 module MarshalDelegateExtensions =
     open System.Collections.Concurrent